little clean-up
