+++ app/src/main/java/com/example/maintainease/data/Converters.kt
@@ -3,6 +3,7 @@ package com.example.maintainease.data
 import androidx.room.TypeConverter
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
+import java.util.Date
 
 class Converters {
     @TypeConverter
@@ -17,4 +18,15 @@ class Converters {
     fun toStringList(list: List<String>?): String? {
         return list?.let { Gson().toJson(it) }
     }
+
+    @TypeConverter
+    fun fromDate(date: Date?): Long? {
+        return date?.time
+    }
+
+    @TypeConverter
+    fun toDate(milliseconds: Long?): Date? {
+        return milliseconds?.let { Date(it) }
+    }
+
 }
